Refactor presentation layer
