+++ app/src/main/java/ru/neexol/rtut/presentation/theme/Colors.kt
@@ -1,0 +1,23 @@
+package ru.neexol.rtut.presentation.theme
+
+import androidx.compose.material.darkColors
+import androidx.compose.ui.graphics.Color
+
+val Purple200 = Color(0xFFBB86FC)
+val Purple700 = Color(0xFF3700B3)
+val Teal200 = Color(0xFF03DAC5)
+
+val RTUtPalette = darkColors(
+	primary = Purple200,
+	primaryVariant = Purple700,
+	secondary = Teal200
+
+/* Other default colors to override
+    background = Color.White,
+    surface = Color.White,
+    onPrimary = Color.White,
+    onSecondary = Color.Black,
+    onBackground = Color.Black,
+    onSurface = Color.Black,
+    */
+)
