+++ app/src/main/java/ru/neexol/rtut/presentation/theme/Typography.kt
@@ -6,7 +6,6 @@ import androidx.compose.ui.text.font.FontFamily
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.sp
 
-// Set of Material typography styles to start with
 val Typography = Typography(
 	body1 = TextStyle(
 		fontFamily = FontFamily.Default,
